requirements.txt
fastapi
uvicorn[standard]
boto3
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
httpx
pytest
pytest-asyncio
pytest-cov
moto[dynamodb,s3,sqs]
localstack-client
mypy
ruff
black